FIX : typage de delta dans animation_rainbow

+++ animations/animation_rainbow/animation_rainbow.cpp
@@ -49,7 +49,7 @@ class AnimationRainbow : public Animation {
 
         int height;
         int width;
-        int delta;
+        double delta;
 
 };
